requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
lxml==5.1.0
pandas==2.1.4
schedule==1.2.1
